java 7 adaptions
