+++ src/plugins/roar/src/resources/i18n/roar_i18n.properties
@@ -32,5 +32,5 @@ roar.display.system = Operating System Notification
 roar.warning.topright = Custom Roar notification style. Fully customizable.
 roar.warning.bottomright = Custom Roar notification style. Fully customizable.
 roar.warning.toaster = Spark notification style. Colors are <u>not</u> customizable.
-roar.warning.system = System-dependent notification. Colors are <u>not</u> customizable.
-roar.warning.system.mac = Uses the Notification-Center. Colors are <u>not</u> customizable.<br>See Mac-Systemsettings -&gt; Notifications for more settings.
+roar.warning.system = System-dependent notification. Colors and duration are <u>not</u> customizable.
+roar.warning.system.mac = Uses the Notification-Center. Colors and duration are <u>not</u> customizable.<br>See Mac-Systemsettings -&gt; Notifications for more settings.
