+++ src/plugins/roar/src/resources/i18n/roar_i18n_de.properties
@@ -32,5 +32,5 @@ roar.display.system = Betriebssystem-Mitteilung
 roar.warning.topright = Roar-spezifische Popups. Flexibel anpassbar.
 roar.warning.bottomright = Roar-spezifische Popups. Flexibel anpassbar.
 roar.warning.toaster = Spark Toaster Popups. Farben sind <u>nicht</u> anpassbar.
-roar.warning.system = Betriebssystemabhängige Popups. Farben sind <u>nicht</u> anpassbar.
-roar.warning.system.mac = Benutzt das Mitteilungs-Center. Farben sind <u>nicht</u> anpassbar.<br>Weitere Einstellungen unter: Mac-Systemeinstellungen -&gt; Mitteilungen.
+roar.warning.system = Betriebssystemabhängige Popups. Farben und Dauer sind <u>nicht</u> anpassbar.
+roar.warning.system.mac = Benutzt das Mitteilungs-Center. Farben und Dauer sind <u>nicht</u> anpassbar.<br>Weitere Einstellungen unter: Mac-Systemeinstellungen -&gt; Mitteilungen.
